chore: lift up label-editing/LabelUtil
