+++ lib/draw/BpmnRenderer.js
@@ -15,7 +15,7 @@ import {
 
 import {
   getLabel
-} from '../features/label-editing/LabelUtil';
+} from '../util/LabelUtil';
 
 import { is } from '../util/ModelUtil';
