+++ lib/features/label-editing/LabelEditingProvider.js
@@ -4,7 +4,7 @@ import {
 
 import {
   getLabel
-} from './LabelUtil';
+} from '../../util/LabelUtil';
 
 import {
   is
