+++ lib/features/label-editing/cmd/UpdateLabelHandler.js
@@ -1,7 +1,7 @@
 import {
   setLabel,
   getLabel
-} from '../LabelUtil';
+} from '../../../util/LabelUtil';
 
 import {
   getExternalLabelMid,
