+++ lib/features/modeling/BpmnUpdater.js
@@ -19,10 +19,10 @@ import {
 import { isAny } from './util/ModelingUtil';
 
 import {
+  getLabel,
   isLabel,
   isLabelExternal
 } from '../../util/LabelUtil';
-import { getLabel } from '../label-editing/LabelUtil';
 
 import { delta } from 'diagram-js/lib/util/PositionUtil';
