+++ lib/features/modeling/behavior/LabelBehavior.js
@@ -11,13 +11,10 @@ import {
 
 import {
   isLabelExternal,
+  getLabel,
   hasExternalLabel,
   isLabel
 } from '../../../util/LabelUtil';
-
-import {
-  getLabel
-} from '../../label-editing/LabelUtil';
 
 import {
   getLabelAdjustment
