+++ lib/features/search/BpmnSearchProvider.js
@@ -6,7 +6,7 @@ import {
 
 import {
   getLabel
-} from '../label-editing/LabelUtil';
+} from '../../util/LabelUtil';
 
 /**
  * @typedef {import('diagram-js/lib/core/Canvas').default} Canvas
