+++ lib/import/BpmnImporter.js
@@ -6,7 +6,8 @@ import { is } from '../util/ModelUtil';
 
 import {
   isLabelExternal,
-  getExternalLabelBounds
+  getExternalLabelBounds,
+  getLabel
 } from '../util/LabelUtil';
 
 import {
@@ -16,10 +17,6 @@ import {
 import {
   isExpanded
 } from '../util/DiUtil';
-
-import {
-  getLabel
-} from '../features/label-editing/LabelUtil';
 
 import {
   elementToString
